refactor: organize crates
